+++ model/table.go
@@ -15,18 +15,18 @@ func (ZMBanner) TableName() string {
 
 // ZMOrder
 type ZMOrder struct {
-	Id        int    `json:"-"`
-	Name      string `json:"name"`
-	OpenId    string `json:"open_id"`
-	UserId    int64  `json:"user_id"`
-	OrderId   int64  `json:"order_id"`
-	Type      int    `json:"type"`
-	CPrice    int    `json:"c_price"`
-	OPrice    int    `json:"o_price"`
-	Number    int    `json:"number"`
-	NumberExt int    `json:"number_ext"`
-	Status    int    `json:"status"`
-	PayTime   string `json:"pay_time"`
+	Id        int     `json:"-"`
+	Name      string  `json:"name"`
+	OpenId    string  `json:"open_id"`
+	UserId    int64   `json:"user_id"`
+	OrderId   int64   `json:"order_id"`
+	Type      int     `json:"type"`
+	CPrice    float64 `json:"c_price"`
+	OPrice    float64 `json:"o_price"`
+	Number    int     `json:"number"`
+	NumberExt int     `json:"number_ext"`
+	Status    int     `json:"status"`
+	PayTime   string  `json:"pay_time"`
 }
 
 // TableName ZMOrder 表名
@@ -36,14 +36,14 @@ func (ZMOrder) TableName() string {
 
 // ZMPay
 type ZMPay struct {
-	Id        int    `json:"id"`
-	Name      string `json:"name"`
-	CPrice    int    `json:"c_price"`
-	OPrice    int    `json:"o_price"`
-	Number    int    `json:"number"`
-	NumberExt int    `json:"number_ext"`
-	Checked   bool   `json:"checked"`
-	Type      int    `json:"-"`
+	Id        int     `json:"id"`
+	Name      string  `json:"name"`
+	CPrice    float64 `json:"c_price"`
+	OPrice    float64 `json:"o_price"`
+	Number    int     `json:"number"`
+	NumberExt int     `json:"number_ext"`
+	Checked   bool    `json:"checked"`
+	Type      int     `json:"-"`
 }
 
 // TableName ZMPay 表名
